Provide additional bundle syntax
Bundle syntax now can support arbitrary arguments on top of
just the initialization options.

This can be used, for instance, for adding in dependencies at
the time of Bundle usage, not just Bundle definition.

For example, you could have a bundle that depends on a web
server, and your parent bundle could pair it with Apache OR nginx.