Add basic execution
Add a SystemTools library that provides wrappers for standard
file system and operating system level functions:

* writeFile
* chmod
* chown
* symlinks
* exec

Also added hook in Executor to loop through the Resources in the queue
and call exec on them with the SystemTools as a single argument.  This
gives each resource the ability to apply itself to the system.