Moving from sync => async
This is the first step in moving from sync to async.
Adopted Q library for promises and uses chains for execution.
At this point none of the actual work is done asynchronously,
but it uses promises for callback flow between resource `exec`
calls.

Things to do next: Update a simple resource to be fully async.